refactor: Use CLI options to generate test coverage
In the next release npm test will run test without generating coverage reports.

For now `npm test` and `npm run test:coverage` will do the same. This is to prevent a breaking change and will be changes with the next major release. Feel free to change this in your project setup.