feat: Auto format code using prettier as pre-commit hook
BREAKING CHANGE:

This commit introduces a new dev dependency which requires Node.JS `>= 8.6`

We are using [prettier](https://prettier.io) to format JavaScript, JSON and SCSS files automatically before you commit your files to Git via a pre-commit hook.

The prettier settings are defined in `.prettierrc` in the project root. In case prettier is to opinated for you or you don’t want Prettier to change your files without the chance to review the changes you just have to delete the pre-commit hook with in the `package.json`:

```json
"husky": {
  "hooks": {
    "post-merge": "npm install",
    "pre-commit": "lint-staged"
  }
}
```

But we totally recommend you to give this workflow a chance, because it’s just one more thing you don’t have to care about.